Introduce BaseCreationConfig
Using ComponentPropertiesImpl as the base type received
by the Task CreationAction classes is not going to work
if we want to have action that can act on AndroidTest
components and App variants (and Dyn-Feature/Test ones).

So we need the CreationAction to act on interfaces instead
where we have more flexibility to have different variants
/components implement more than one.

For now just introduce the base interface BaseCreationConfig
and haveComponentPropertiesImpl implement it, and rename
the method param to creationConfig.

Bug: N/A
Test: existing tests
Change-Id: Ia0835872f8d2b458d6df8ffe36e4eba1746fbb24