Always return Status::OK in GetCpuCoreConfig command
To be consistent with Transport gRPC pattern, we should always
return Stats::OK for non-fatal errors in commands.

GetCpuCoreConfig always fails on emulators but we have a fallback
on the host side so there's no need to surface this failure as a
gRPC error. Instead we'll just log a warning message in logcat.
This will clean up error handling on the host side.

Bug: N/A
Test: manually verify warning message
Change-Id: I94df5fe1fcb993a6398b23543070de0fd00d2096