Convert auto namespacing to an artifact transform.
This includes generating the namespaced R.jar from the symbol list with
package name to enable those code paths to be merged, and to stop
shipping the R.jar in the AAR.

This change means that the R class is fully compile classpath scoped
when namespacing is enabled, so fixes up various tests to do with that
too.

Dynamic feature support is close, but needs more work around R class
handling and the bundle.

Bug: 69358522
Bug: 72695265
Bug: 146340124
Fixes: 110879504
Fixes: 111153788
Fixes: 111974213
Fixes: 128825217
Fixes: 138780160
Fixes: 141887465
Test: Covered by existing AGP integration tests
Test: Added some new unit tests.
Change-Id: I97d887d03da0008286ea594bccf8c3fff380887b