Guarantee orders of the stages of Task Creation.
When running register on a task container that has callback
using the non lazy APIs, then Action.execute is called during
the register before we get access to the TaskProvider.

Since we use a wrapper we can delay calling configure()
on our TaskCreationAction until after the register method
has returned, and after we first called handleProvider.

This delay has no impact because current callbacks
the current non-lazy APIs do not guarantee that the
callbacks are called after the task configuration
actions. In fact they seem to always happen before!

Bug: N/A
Test: new unit tests
Change-Id: Iabee5e96c35ade250c5bd54871bb630b6c2c9cce